Entity::update() virtual (old: pure virtual)
Entities do not require to be updatable anymore, it is optional.
Some entities might not be updatable at all, and some might require
more parameters to be updated. The generic Entity::update() call is
kept in order to be able to iterate and update some simple entities
(e.g. lights).